added customer operation

+++ src/main/java/com/aselsis/aselmanager/dto/SaveOrderDto.java
@@ -5,6 +5,7 @@ import lombok.Data;
 
 @Data
 public class SaveOrderDto {
-    private Integer totalPrice;
+
+    private Double totalPrice;
 
 }
